protected ui and routes

+++ src/features/ticket/actions/delete-ticket.ts
@@ -1,14 +1,31 @@
 "use server";
 
 import { setCookieByKey } from "@/actions/cookies";
-import { fromErrorToActionState } from "@/components/forms/utils/to-action-state";
+import {
+  fromErrorToActionState,
+  toActionState,
+} from "@/components/forms/utils/to-action-state";
+import { getAuthOrRedirect } from "@/features/auth/queries/get-auth-or-redirect";
+import { isOwner } from "@/features/auth/utils/is-owner";
 import { prisma } from "@/lib/prisma";
 import { ticketsPath } from "@/paths";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 
 export async function deleteTicket(id: string) {
+  const { user } = await getAuthOrRedirect();
+
   try {
+    const ticket = await prisma.ticket.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!ticket || !isOwner(user, ticket)) {
+      return toActionState("ERROR", "Not authorized");
+    }
+
     await prisma.ticket.delete({
       where: {
         id,
